Replace global $_ variable with lexical
In _xor_mask function, the $_ variable is used to store temporary data.
Unfortunately $_ is not protected with local. In some case, this lead to
tht following error (on Debian's mojolicious 2.96):

EV: error in callback (ignoring): Modification of a read-only value attempted at /usr/share/perl5/Mojo/Transaction/WebSocket.pm line 301.

Changing $- to a lexical variable ($snip in this case) fixes the issue.